feat(chatbot): 重构聊天记录处理并添加流式响应支持- 将 Chat 结构改为指针类型，优化内存使用 - 添加新消息标记，用于优化聊天记录加载 - 实现流式响应处理，提升用户体验 -优化 API 请求逻辑，增加错误处理和日志记录 - 删除不必要的 GetBilling 函数，简化代码结构

+++ include/Logger.h
@@ -19,7 +19,7 @@ public:
         return std::chrono::duration_cast<std::chrono::milliseconds>(duration).count();
     }
 
-    static std::string Stamp2Time(long long int timestamp, bool nospace = false) {
+    static std::string Stamp2Time(size_t timestamp, bool nospace = false) {
         int ms = timestamp % 1000;//取毫秒
         time_t tick = (time_t) (timestamp / 1000);//转换时间
         struct tm tm;
